rollodex: Add initial version
Adds an initial version of the "rollodex", a server that facilitates
members of a mesh network being able to talk to each other. The rollodex
serves two purposes:

- Opening a method to network members behind a NAT
- Communicating the NAT holes to other members

The idea is that members can use this to bootstrap communications in the
mesh. No sensitive traffic gets sent to the rollodex.