rollodex: Tell network members their own IP
Adds an extra field in the message sent from the rollodex to inform the
network members what their own IP address is. This is helpful to stop
them trying to connect to themselves and also allows them to figure out
if they should initiate contact over the link or wait for it.