mesh member: Add first cut of class
MeshMember encapsulates a member of the mesh network. It:

- Manages a connection to the rollodex, so that it can expose its UDP
  port and hear about new members connecting to the mesh
- Connects to other members
- Allows other member to connect to it

A few notes about how this currently works:

- The channel between members is dtls based on a PSK
- Because dtls needs the concept of a "server" and "client" the members
  choose this (arbitrarily) by setting the member with the higher
  IP/Port to be the server and the other to be the client
- The rollodex only tells members of a mesh about the outside (non VPN)
  addresses of other members in the mesh. Members need to connect to
  each other (via the DTLS channels) in order to learn their inside VPN
  addresses.
- The inside IP address (VPN address) is transmitted via the dtls "hint"
  message as this can be relied upon to be received. In this way I can
  rely on this being sent and received over UDP and don't need to come
  up with my own handshake protocol.

This is currently largely unfinished but is currently kind of working.
It currently is a bit hokey and connections between members have issues
starting up a lot of the time. I suspect this is because the "server"
member starts listening right as the "client" member joins. This can be
made a bit nicer by only starting the server when the client reaches
out.