rollodex: Avoids occasional deadlock in register
The register function locks the member map and then sends to a channel
to let the serve routine know to send out a new update. If the serve
channel is already in the process of sending out an update then writing
to the channel can block, which means the register function retains the
membersmap lock and the two routines are deadlocked. To avoid, release
the lock before notifying via the channel.