peer connector: Open 'server' peer's firewall
Before this change, the (arbitrarily chosen) dTLS server peer only
awaited connection from the peer chosen to be the client. If the server
peer is behind a stateful firewall, this would mean that no traffic
would be able to get through to it and the connection wouldn't be able
to be formed.

This change makes the server peer send a message outbound to the peer,
when it learns about the peer in order to open up its firewall such that
the peer can then communicate with it.